[refactor] #47 온보딩 관련 코드 복구
