+++ app/src/main/java/com/dpm/sixpack/main/navigation/MainNavHost.kt
@@ -8,7 +8,7 @@ import androidx.navigation.compose.NavHost
 import androidx.navigation.navOptions
 import com.dpm.sixpack.SixPackAppState
 import com.dpm.sixpack.presentation.destinations.OnboardingRoute
-import com.dpm.sixpack.presentation.routes.deprecated.onboarding.navigation.addOnboardingNavGraph
+import com.dpm.sixpack.presentation.routes.onboarding.navigation.addOnboardingNavGraph
 import com.dpm.sixpack.presentation.routes.session.navigation.addRunningSessionNavGraph
 import com.dpm.sixpack.presentation.routes.sessionreport.navigation.addSessionReportNavGraph
