--- presentation/src/main/java/com/dpm/sixpack/presentation/routes/deprecated/onboarding/contract/uistate/level/RunningLevelConstants.kt
@@ -1,7 +1,0 @@
-package com.dpm.sixpack.presentation.routes.deprecated.onboarding.contract.uistate.level
-
-object RunningLevelConstants {
-    const val BEGINNER = "BEGINNER"
-    const val OCCASIONAL = "OCCASIONAL"
-    const val CONSISTENT = "CONSISTENT"
-}
