+++ presentation/src/main/java/com/dpm/sixpack/presentation/routes/onboarding/contract/uistate/level/RunningLevelConstants.kt
@@ -1,0 +1,7 @@
+package com.dpm.sixpack.presentation.routes.onboarding.contract.uistate.level
+
+object RunningLevelConstants {
+    const val BEGINNER = "BEGINNER"
+    const val OCCASIONAL = "OCCASIONAL"
+    const val CONSISTENT = "CONSISTENT"
+}
