+++ presentation/src/main/java/com/dpm/sixpack/presentation/routes/onboarding/contract/uistate/permission/TermType.kt
@@ -1,0 +1,14 @@
+package com.dpm.sixpack.presentation.routes.onboarding.contract.uistate.permission
+
+import androidx.annotation.StringRes
+import com.dpm.sixpack.presentation.R
+
+enum class TermType(
+    @StringRes val title: Int,
+    val isRequired: Boolean,
+    @StringRes val description: Int? = null,
+) {
+    LOCATION(R.string.onboarding_permission_term_location, true),
+    PRIVACY(R.string.onboarding_permission_term_privacy, true),
+    MARKETING(R.string.onboarding_permission_term_marketing, false),
+}
