+++ presentation/src/main/java/com/dpm/sixpack/presentation/routes/onboarding/ui/component/level/LevelCardList.kt
@@ -1,0 +1,51 @@
+package com.dpm.sixpack.presentation.routes.onboarding.ui.component.level
+
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.height
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+import com.dpm.sixpack.presentation.common.components.preview.DoRunPreviewWrapper
+import com.dpm.sixpack.presentation.routes.onboarding.contract.uistate.level.LevelType
+
+@Composable
+fun LevelCardList(
+    modifier: Modifier = Modifier,
+    selectedLevel: LevelType? = null,
+    onSelectLevel: (LevelType) -> Unit = {},
+) {
+    LevelType.entries.forEach { level ->
+        LevelCard(
+            level = level,
+            isSelected = level == selectedLevel,
+            onSelectLevel = onSelectLevel,
+            modifier = modifier,
+        )
+
+        Spacer(Modifier.height(16.dp))
+    }
+}
+
+@Preview
+@Composable
+private fun LevelCardListPreview1() {
+    DoRunPreviewWrapper {
+        Column(modifier = Modifier) {
+            LevelCardList(
+                selectedLevel = LevelType.BEGINNER,
+            )
+        }
+    }
+}
+
+@Preview
+@Composable
+private fun LevelCardListPreview2() {
+    DoRunPreviewWrapper {
+        Column(modifier = Modifier) {
+            LevelCardList()
+        }
+    }
+}
